[presets] add new module containing the applyPresets function from videoproof-contextual/models, also use in type-tools-grid.
