+++ .prettierignore
@@ -22,6 +22,7 @@
 !lib/js/**/*.css
 !**/ramp/*.*
 !lib/js/components/**/videoproof-contextual/*
+!lib/js/components/presets.mjs
 
 # Ignore glyph-groups.yaml
 lib/assets/glyph-groups.yaml
