+++ lib/js/components/actors/videoproof-contextual/models.mjs
@@ -25,10 +25,6 @@ import {
     createAvailableTypes,
     createDynamicType,
     StaticDependency,
-    unwrapPotentialWriteProxy,
-    deserializeSync,
-    SERIALIZE_OPTIONS,
-    SERIALIZE_FORMAT_OBJECT,
 } from "../../../metamodel.mjs";
 
 import {
@@ -43,6 +39,8 @@ import { _BaseActorModel, genericActorMixin } from "../actors-base.mjs";
 import { ColorModel } from "../../color.mjs";
 
 import { CharGroupModel } from "../videoproof-array.mjs";
+
+import { applyPresets, CUSTOM_PRESET_KEY } from "../../presets.mjs";
 
 // Coherence: ensure charGroups has 1-2 items
 function ensureAtLeastOneCharGroup({ charGroups }) {
@@ -196,8 +194,7 @@ export const TemplateRuleModel = _AbstractStructModel.createClass(
 //   - kern-upper: default 'HO\1\2\1OLA', \1=Latin.Uppercase, \2=same
 //   - kern-mixed: default '\1\2nnoy', \1=Latin.Uppercase \2=Latin.Lowercase,
 //   - kern-lower: default 'no\1\2\1ony', \1r=Latin.Lowercase, \2=same
-const CUSTOM_PRESET_KEY = "custom",
-    PRESETS_OPTIONS = new Map([
+const PRESETS_OPTIONS = new Map([
         [
             "auto-short",
             {
@@ -303,142 +300,42 @@ const CUSTOM_PRESET_KEY = "custom",
         ],
         [CUSTOM_PRESET_KEY, null],
     ]),
-    _presets_options_keys = Array.from(PRESETS_OPTIONS.keys()),
-    ContextualPresetModel = _AbstractEnumModel.createClass(
+    _presets_options_keys = Array.from(PRESETS_OPTIONS.keys());
+export const ContextualPresetModel = _AbstractEnumModel.createClass(
         "ContextualPresetModel",
         _presets_options_keys,
         _presets_options_keys.at(0),
     ),
-    GENERATED_DATA = Symbol.for("GENERATED_DATA");
-
-function copyToDraft(item) {
-    const draftItem = item.constructor.createPrimalDraft(item.dependencies);
-    for (const [key, entry] of item) {
-        if (item instanceof _AbstractListModel) draftItem.push(entry);
-        else draftItem.set(key, entry);
-    }
-    return draftItem;
-}
-
-/**
- * By injecting PRESETS_OPTIONS and CUSTOM_PRESET_KEY I hope this can
- * be used as a pattern for other preset implementations as well!
- */
-function applyPresets(
-    presetsFieldName,
-    PRESETS_OPTIONS,
-    CUSTOM_PRESET_KEY,
-    valuesMap,
-    { isNew /*, wasSerialized*/, setFn },
-) {
-    const {
-        [presetsFieldName]: presets,
-        ...settables // {charGroups, template}}
-    } = valuesMap;
-
-    // - If any item of settables is customized the presets.value
-    //   should become CUSTOM_PRESET_KEY ("custom") as well.
-    // - When the type has changed to a not "custom" key, the
-    //  settables must be changed as well.
-    //
-    // The idea is that the GENERATED_DATA symbol will
-    // disappear when edited outside of this function
-    // and hence we would set the type to "custom"
-
-    const presetsChanged = !!unwrapPotentialWriteProxy(presets, "draft");
-
-    let apply = "";
-    if (presets.value !== CUSTOM_PRESET_KEY && (isNew || presetsChanged)) {
-        // isNew: wouldn't have the data applied
-        // presetsChanged: whatever the data is now, it will be replaced
-        apply = "preset";
-    } else if (presets.value === CUSTOM_PRESET_KEY && presetsChanged) {
-        apply = CUSTOM_PRESET_KEY;
-    } else {
-        const presetData = PRESETS_OPTIONS.get(presets.value);
-        for (const [key, entry] of Object.entries(settables)) {
-            // depending on the presence in the presets we skip the entry here!
-            // i.e. if charGroups is settable, but the current preset doesn't
-            // set it, it can't trigger apply(CUSTOM_PRESET_KEY).
-            if (presetData && !(key in presetData)) continue;
-            const unwrapped = unwrapPotentialWriteProxy(entry);
-            if (!Object.hasOwn(unwrapped, GENERATED_DATA)) {
-                // entry has been customized/not generated/created in here.
-                apply = CUSTOM_PRESET_KEY;
-                presets.value = CUSTOM_PRESET_KEY;
-                break;
-            }
-        }
-    }
-    // apply a state if required...
-    if (apply === "preset") {
-        const serializeOptions = Object.assign({}, SERIALIZE_OPTIONS, {
-                format: SERIALIZE_FORMAT_OBJECT,
-            }),
-            presetData = PRESETS_OPTIONS.get(presets.value);
-        for (const [key, entry] of Object.entries(settables)) {
-            if (!(key in presetData)) continue;
-            const Ctor = unwrapPotentialWriteProxy(entry).constructor,
-                serializedData = presetData[key],
-                newEntry = deserializeSync(
-                    Ctor,
-                    entry.dependencies,
-                    serializedData,
-                    serializeOptions,
-                );
-            // the new Entry is immutable and we need it to
-            // be a draft with actual changes so the GENERATED_DATA
-            // flag sticks to it...
-            const entryDraft = copyToDraft(newEntry);
-            // Actually the mere presence of this GENERATED_DATA
-            // symbol is enough, the value may be helping when debugging.
-            entryDraft[GENERATED_DATA] = presets.value;
-            setFn(key, entryDraft);
-        }
-    } else if (apply === CUSTOM_PRESET_KEY) {
-        // The data will have to be serialized.
-        // Make sure the drafts are going to be new items
-        // and not reverted back, so that the GENERATED_DATA
-        // symbol cannot stick! Create a copy:
-        for (const [key, entry] of Object.entries(settables)) {
-            const newEntry = copyToDraft(entry); // => new item
-            // this should have deleted any trace of GENERATED_DATA
-            setFn(key, newEntry);
-        }
-    }
-}
-
-export const VideoproofContextualKeyMomentModel =
-        _AbstractStructModel.createClass(
-            "VideoproofContextualKeyMomentModel",
-            ...typographyKeyMomentModelMixin,
-            // TODO: when we are going for dynamically loaded, user-defined
-            // presets this may have to change. It could still be driven
-            // by that enum, e.g. adding a "user" option, but the value
-            // for the user defined option would be a string database key
-            // or so.
-            ["presets", ContextualPresetModel],
-            ["charGroups", CharGroupArgumentsListModel],
-            ["template", TemplateModel],
-            ["stageBackgroundColor", ColorModel],
-            CoherenceFunction.create(
-                ["presets", "charGroups", "template"],
-                applyPresets.bind(
-                    null,
-                    "presets",
-                    PRESETS_OPTIONS,
-                    CUSTOM_PRESET_KEY,
-                ),
-            ),
-            // These run after the presets to ensure **coherence**.
-            CoherenceFunction.create(["charGroups"], ensureAtLeastOneCharGroup),
-            CoherenceFunction.create(
-                ["charGroups"],
-                function capCharGroupsAtTwo({ charGroups }) {
-                    if (charGroups.size > 2) charGroups.splice(2, Infinity);
-                },
+    VideoproofContextualKeyMomentModel = _AbstractStructModel.createClass(
+        "VideoproofContextualKeyMomentModel",
+        ...typographyKeyMomentModelMixin,
+        // TODO: when we are going for dynamically loaded, user-defined
+        // presets this may have to change. It could still be driven
+        // by that enum, e.g. adding a "user" option, but the value
+        // for the user defined option would be a string database key
+        // or so.
+        ["presets", ContextualPresetModel],
+        ["charGroups", CharGroupArgumentsListModel],
+        ["template", TemplateModel],
+        ["stageBackgroundColor", ColorModel],
+        CoherenceFunction.create(
+            ["presets", "charGroups", "template"],
+            applyPresets.bind(
+                null,
+                "presets",
+                PRESETS_OPTIONS,
+                CUSTOM_PRESET_KEY,
             ),
         ),
+        // These run after the presets to ensure **coherence**.
+        CoherenceFunction.create(["charGroups"], ensureAtLeastOneCharGroup),
+        CoherenceFunction.create(
+            ["charGroups"],
+            function capCharGroupsAtTwo({ charGroups }) {
+                if (charGroups.size > 2) charGroups.splice(2, Infinity);
+            },
+        ),
+    ),
     VideoproofContextualKeyMomentsModel = _AbstractListModel.createClass(
         "VideoproofContextualKeyMomentsModel",
         VideoproofContextualKeyMomentModel,
