+++ lib/js/components/layouts/type-tools-grid.mjs
@@ -11,10 +11,6 @@ import {
   , InternalizedDependency
   , BooleanDefaultTrueModel
   , CoherenceFunction
-  , deserializeSync
-  , SERIALIZE_OPTIONS
-  , SERIALIZE_FORMAT_OBJECT
-  , unwrapPotentialWriteProxy
 } from '../../metamodel.mjs';
 
 import {
@@ -183,6 +179,11 @@ import {
     ProseMirror
 } from '../prosemirror/integration.typeroof.jsx';
 
+import {
+    CUSTOM_PRESET_KEY
+  , applyPresets
+} from '../presets.mjs';
+
 import { schemaSpec as proseMirrorDefaultSchema } from "../prosemirror/very-simple-schema";
 
     /**
@@ -226,14 +227,29 @@ const GRID_CONTENT_OPTIONS = new Map([
         ['A-Z', 'ABCDEFGHIJKLMNOPQRSTUVWXYZ']
       , ['a-z', 'abcdefghijklmnopqrstuvwxyz']
       , ['0-9', '0123456789']
-      , ['custom', null]
+      , [CUSTOM_PRESET_KEY, null]
       // I'd like to load a custom value for this from the registered-properties
       // type: 'custom' with a customText: 'H'
-    ])
+    ].map(([key, textContent])=>{
+        const documentData = {
+            "typeKey": "doc",
+            "content": [
+                {
+                    "typeKey": "paragraph",
+                    "content": [
+                        {
+                            "typeKey": "text",
+                            "text": textContent
+                        }
+                    ]
+                },
+            ]
+        };
+        return [key, {document: documentData}];
+    }))
   , _grid_content_options_keys = Array.from(GRID_CONTENT_OPTIONS.keys())
   , GridContentTypeModel = _AbstractEnumModel.createClass('GridContentTypeModel',
         _grid_content_options_keys, _grid_content_options_keys.at(0))
-  , GENERATED_DATA = Symbol.for('GENERATED_DATA')
   , GridContentModel = _AbstractStructModel.createClass(
          'GridContentModel'
        , ['type', GridContentTypeModel]
@@ -246,85 +262,16 @@ const GRID_CONTENT_OPTIONS = new Map([
        // a flag, the content could be useful for debugging maybe.
        // In a second step the data is skipped in serialization, so we
        // re-generate it on load.
-       //
        , ['document', NodeModel]
        , CoherenceFunction.create(
-        [
-            "type",
-            "document",
-        ],
-        function initDocument({
-            type,
-            document,
-        }) {
-            // If the document is custom the type should become "custom" as well
-            // when the type has changed to a not "custom" key, the document
-            // content must be changed as well.
-
-            // Actually "type has a draft", value may still be the same, but the
-            // presence of a draft is enough so far.
-            const typeHasChanged = !!unwrapPotentialWriteProxy(type, 'draft')
-                // For a ProseMirror document, this seems to be sufficient,
-                // otherwise may be hard to detect!
-              , isNewDocument = document.get("content").size === 0
-              , contentType = type.value
-              ;
-            if (contentType !== 'custom' && (isNewDocument || typeHasChanged)) {
-                const data = {
-                        "typeKey": "doc",
-                        "content": [
-                            {
-                                "typeKey": "paragraph",
-                                "content": [
-                                    {
-                                        "typeKey": "text",
-                                        "text": GRID_CONTENT_OPTIONS.get(contentType)
-                                    }
-                                ]
-                            },
-                        ]
-                    },
-                    serializeOptions = Object.assign(
-                        {},
-                        SERIALIZE_OPTIONS,
-                        {
-                            format: SERIALIZE_FORMAT_OBJECT,
-                        },
-                    ),
-                    newItem = deserializeSync(
-                        NodeModel,
-                        document.dependencies,
-                        data,
-                        serializeOptions,
-                    );
-                for (const [key, enrty] of newItem.entries())
-                    document.set(key, enrty);
-                // The idea is that the symbol will disappear when edited
-                // outside of this function and hence we would set the
-                // type to "custom"
-
-                // documentDraft will be undefined if it is not a draft yet
-                const documentDraft = unwrapPotentialWriteProxy(document, 'draft');
-                // Actually the mere presence of this GENERATED_DATA
-                // symbol is enough, the value may be helping when debugging.
-                documentDraft[GENERATED_DATA] = contentType;
-            }
-            else if (contentType === 'custom' && typeHasChanged) {
-                // The document data will have to be serialized.
-                // Make sure the draft is going to be a new item
-                // and not reverted back, so that the GENERATED_DATA symbol
-                // cannot stick! Create a copy:
-                const content = document.get('content')
-                  , newContent = NodeModel.fields.get('content').createPrimalDraft(content.dependencies)
-                  ;
-                newContent.push(...content.value);
-                document.set('content', newContent);
-            }
-            else if(!Object.hasOwn(unwrapPotentialWriteProxy(document), GENERATED_DATA)) {
-                // document has been customized/not generate/created in here.
-                type.value = 'custom';
-            }
-        })
+            ["type", "document"],
+            applyPresets.bind(
+                null,
+                "type",
+                GRID_CONTENT_OPTIONS,
+                CUSTOM_PRESET_KEY,
+            ),
+        ),
     )
   , GridPropertiesModel = _AbstractStructModel.createClass(
         'GridPropertiesModel'
